+++ docs/conf.py
@@ -12,7 +12,11 @@
 # serve to show the default value.
 
 import datetime
-import pkg_resources
+
+try:
+    import importlib.metadata as importlib_metadata
+except ImportError:
+    import importlib_metadata
 import pylons_sphinx_themes
 
 # If your extensions are in another directory, add it here.
@@ -51,7 +55,7 @@ copyright = '2011-%s, Ian Bicking and contributors' % thisyear
 # other places throughout the built documents.
 #
 # The short X.Y version.
-version = pkg_resources.get_distribution('pastedeploy').version
+version = importlib_metadata.distribution('pastedeploy').version
 
 # The full version, including alpha/beta/rc tags.
 release = version